refactor(admins, threads, users): reorganize modules structure to match auth module
