+++ src/users/controllers/users.controller.spec.ts
@@ -1,6 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { UsersController } from './users.controller';
-import { UsersService } from './services/user.service';
 
 describe('UsersController', () => {
   let controller: UsersController;
@@ -8,7 +7,7 @@ describe('UsersController', () => {
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       controllers: [UsersController],
-      providers: [UsersService],
+      // providers: [UsersService],
     }).compile();
 
     controller = module.get<UsersController>(UsersController);
