+++ src/users/users.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { UserService } from './services/user.service';
-import { UsersController } from './users.controller';
+import { UsersController } from './controllers/users.controller';
 import { ThreadsModule } from '@/threads/threads.module';
 import { UserContentService } from './services/user-content.service';
 import { UserDetailsService } from './services/user-details.service';
